Service Network checkbox can be ticked
The network configuration of an environment, can contain a property
'is_service_network' which will according to its value check or
uncheck the service network checkbox.

If 'is_service_network' is false or absent, the service network
checkbox will not be checked.

If 'is_service_network' is true, the checkbox will be checked.